Update model.py
Fixing the error:
```
Expected object of type Variable[torch.cuda.FloatTensor] but found type Variable[torch.cuda.LongTensor] for argument #1 ‘target’
```
for the rossman notebook. I don't know the impact of that change to the other notebooks nor do I know why it was casted to `long` in the first place. Maybe Jeremy could shed some lights on this.