Correct property typo (self.data.path)
I got an error when I ran the tutorial code of deep learning 1. It said that self.data_path was not defined, so I looked into learner.py and saw that on the line before this line, there was self.data.path used instead of self.data_path, so I believe that should be corrected.