Inconsistent naming of itos.pkl
`pretrain_lm.py` saves the index file as `itos.pkl` in the appropriate tmp folder. But `finetune_lm.py` currently expects `itos_{pretrain_id}.pkl`, so this breaks. Since the location of the tmp file already unambiguously refers to the pre-trained model (`pretrain_path / tmp / itos.pkl`). Would it be more consistent to just expect `itos.pkl` (as in this change)? Or was it done this way because current users tend to have `dir_path == pretrain_path`? Given that they are separate parameters, I thought this change was more sensible.