Update image.py
The function open_mask would return a tensor with the wrong shape [4, height, width] if the pixel values in the original mask image are 0 and 255 instead of 0 and 1. Converting the PIL image to grayscale ('.convert('L')) before running it through pil2tensor fixes this. 
This is also similar to the 'open_image' function which uses .convert('RGB').