Add two test of qrnn one that shows the issue desc. in #1108
The issue appears only if one does not use split_fn when language model is contructed.

As described on the forum thread: https://forums.fast.ai/t/multilingual-ulmfit/28117/21 the issue is caused by  WeightDropout that is not intialised properly, and then wrong ( not sure how yet) paramters are passed to the optymiser and only part of the network gets gradinet changes and rest stays at random.