Fix backpropagation of LSTM when trained on fp16 (#1374)
Otherwise you get an error that the Backpropagation mixes fp16 with fp32. loss is fp32 while mean of weights is fp16. The proposed solution simply converts AR & TAR to fp32 always. This should be a noop in normal training and it fixes the issue during fp16 training.

I don't have a card that supports fp16 temporarly avaliable so I can't test the changes.