Change formatting of CSV path
The assignment of csv_path was generating an error on certain setups. The line:

csv_path = self._path/'cleaned.csv'

was generating the error:

TypeError: unsupported operand type(s) for /: 'str' and 'str'

So it thought that the / was an operand for some reason. Instead I  used os.path.join, which is much more universally compatible than the code used before.