Make classes to labels consistent
Right now when you get labels from_csv or df it infers the type meaning if I get the labels from a CSV file and the labels are all numeric values it will assume they are an integer if not it will assume they are strings. However, label_from_df always assume the labels are string even if all the folder names are numeric. In isolation, this those not cause any problem but if I try to create a DataBunch where I read the train is from a csv file and the validation is from folder there will be an inconsistency between the mapping of classes names and labels. That is because sorting numeric string is different than sorting integers. 

I know my case is very unique but I experienced this and others might 2. The idea is that both labels_from_df and label_from_folder provide the same mapping which they currently don't mainly because string sorted not the same as int sorting when creating the label name to label dictionary.