Proposal for model_summary update
I did a bit of work on model.summary() as discussed in https://forums.fast.ai/t/developer-chat/22363/942?u=sebastienwood
Now the model name, optimizer, loss function and callbacks functions are listed as well. Extra code for AdamW. Would however require to further discuss what and whatnot include, and how to present it to the users.