Add support for training sampler
This adds support for a sampler to use for the training set DataLoader.
This was motivated by wanting to use a WeightedRandomSampler to support
oversampling of rare classes. The sampler can't simply be passed in as a
kwarg to databunch() because we only want to pass it to the train_dl.
Also, when specifying a sampler, we need to turn off shuffle for the
train_dl because that option is not compatible with setting a sampler --
PyTorch won't allow it.