Bug Fix: `plot_top_losses`
There was a bug happening where since they had the same name, tabular plot_top_losses was taking priority to vision's plot_top_losses. To adjust for this I mimicked how it is for the `plot_multi_top_losses` and just renamed the function to `plot_tab_top_losses`. I had originally thought that this would not have been an issue at the time of making the original PR but since they all have the same super interp class I think the functions were at conflict of which to use sometimes. This resolves the issue by just including a third plot option for tabular.