Update awd_lstm.py (#2517)
When the weights in an embedding layer are initialised all of the weights are updated, including those set to zero at the padding index. This means that the weights at the padding idx are now non zero, which I don't think can be desirable. The change I am proposing just reinstates the zeros at the padding index.