Fix potential bug in 04_data.external
In URLs.path(), there is a check to see if the file to download
exists in folders `data` or `models` in current working directory.

The key to the `Config` is set to `models` instead of `model` when
forming the local path. The key shoule be one of ['archive', 'data',
'model', 'storage'] according to the docs