_array2tensor() now accepts **kwargs
The aforementioned function did not previously accept **kwargs even though it was being passed to it in line 152 which caused it to throw an error when only a single float was passed to torch_core.tensor() as mentioned in issue #3781. This commit adds **kwargs to _array2tensor() and uses it to set the requires_grad attr of the returned tensor from the value that was passed using **kwargs by user.