Update the fix to be a non-breaking change
The issue with the previous code is that if someone was initializing a
DistributedDL and required that the DistributedDL be on the same device
as the DataLoader that it's wrapping, then the old code would have
broken that functionality.

This commit makes it so that we only muck with the previous behavior in
the case where someone explicitly passes in `device` as a kwarg,
effectively _opting in_ to the new functionality.